requirements.txt
cairocffi==0.9.0
CairoSVG==2.2.1
cffi==1.11.5
cssselect2==0.2.1
defusedxml==0.5.0
Django==2.1.2
django-appconf==1.0.2
django-autocomplete-light==3.3.2
django-braces==1.13.0
django-breadcrumbs==1.1.3
django-clever-selects==0.8.2
django-crispy-forms==1.7.2
django-filter==2.0.0
django-material-widgets==1.0.0b3
django-materializecss-form==1.1.10
django-select2==6.3.1
html5lib==1.0.1
pdfrw==0.4
Pillow==5.3.0
psycopg2==2.7.6.1
psycopg2-binary==2.7.6.1
pycparser==2.19
Pyphen==0.9.5
pytz==2018.6
reportlab==3.5.12
six==1.12.0
tinycss2==0.6.1
WeasyPrint==0.42.3
webencodings==0.5.1
xlrd==1.1.0
